Add files via upload
the purpose of this app created is to help users decide what to eat when they are hungry. this app created randomizes and suggests what the user can eat for breakfast, lunch and dinner. if the user does not like the food given to them, they can simply press the randomize button again until satisfied